cloudrun/ingestion/app/requirements.txt
Flask
python-dateutil
Werkzeug
six
gunicorn
requests
pyarrow
google-cloud
google-cloud-bigquery
google-cloud-core
google-cloud-pubsub